Move parsing methods to calls_parsing.py (#418)
Moves 4 methods to the new CallsParsingMixin:
- parse_nat, parse_int, parse_bool, parse_float64

All emit state-machine loops in WAT that allocate Result<T, String>
on the heap with interned error messages from string_pool.

calls.py: 6,124 -> 5,168 lines (-956)
calls_parsing.py: 15 -> 971 lines
